uploaded France400.AIW
This version has a number of problems. The corridors of the track and pit lane overlap, causing the formation lap to never end. There are also a bunch of waypoint connections that are probably wrong. The pit and garage spots also need to be redone because sometimes cars crash into each other.