Add files via upload
This contains a big pit wall that covers the entire infield except the garage and safety car pit sections, as well as a more accurate wall for the safety car pit and a pole with a better location. It also has an XFINISH object that has been modified to work with these adjustments, as well as XSECTOR1 and XSECTOR2 objects that have had unnecessary planes removed.